requirements.txt
tensorflow
python-mnist
